尝试解耦文字识别（ocr）与屏幕捕获（capture）
